+++ src/database/__init__.py
@@ -1,1 +1,4 @@
+from . import models
+from .setup import initialise_mongodb_connection
 
+__all__ = ["initialise_mongodb_connection", "models"]
