+++ src/database/models/__init__.py
@@ -10,3 +10,19 @@ from .user import (
     User,
     Votes,
 )
+
+__all__ = [
+    "User",
+    "Profile",
+    "Preference",
+    "Record",
+    "Server",
+    "Submissions",
+    "Stats",
+    "Votes",
+    "Channels",
+    "WinCount",
+    "LanguageProblemCount",
+    "SkillProblemCount",
+    "SkillsProblemCount",
+]
