+++ src/ui/embeds/leaderboards.py
@@ -1,9 +1,9 @@
 import discord
-from datetime import datetime, UTC
 import pytz
-from constants import DifficultyScore
-from database.models import Server
-from ui.embeds.common import failure_embed
+
+from src.constants import DifficultyScore
+from src.database.models import Server
+from src.ui.embeds.common import failure_embed
 
 
 def empty_leaderboard_embed() -> discord.Embed:
